fix(hardening): harden audit auth server and wal paths
Bind encrypted audit entries to the hash chain with GCM AAD, enforce configured password policy, make lifecycle drain cancellable, avoid SQL comment false positives, validate encrypted WAL batch sizes, and refresh feature status.

+++ pkg/server/server.go
@@ -92,7 +92,7 @@ func generateRandomPassword() (string, error) {
 		return "", fmt.Errorf("crypto/rand failed: %w", err)
 	}
 	for i := range b {
-		b[i] = charset[b[i]%byte(len(charset))]
+		b[i] = charset[uint32(b[i])%uint32(len(charset))]
 	}
 	return string(b), nil
 }
